chore: support generating armv7l .deb packages
This is the architecture of the Raspberry Pi 3. The following changes
were necessary:

- Detect the architecture using `uname -m`
- Set the Debian equivalent architecture to `armhf`
- Handle the new architecture correctly on `dependencies-npm.sh` and
  `electron-download-package.sh`, referring to it as simply `arm` (which
  is what `node-gyp` expects)

For simplicity purposes, the custom Debian architecture logic in
`electron-installer-debian-linux.sh` was removed by making the script
accept the Debian architecture we're already declaring in the Makefile.

This PR also includes `architecture-get.sh`, which is a script to find
an architecture equivalence for Electron and Debian.

Steps to test this PR:

- `make electron-installer-debian` from a Raspberry Pi 3

Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>